Upgrade SvelteKit (#886)
* Upgrade SvelteKit

* Move vite config to separate file (new SvelteKit requirement)

* Add TODO

* Replace references to port 3000 with port 5173 (new SvelteKit port)

* Rename route template pages (new SvelteKit format)

* Use `$app/environment` not `$app/env` (new SvelteKit path)

* Fix build error

* Use `localhost:3000` not `127.0.0.1:5173`

* Remove unused script (it hasn't been working)

* Enforce use of port 3000 (otherwise CORS errors)

* Add `vite.config.js` to directory in Docker
