+++ package.json
@@ -4,21 +4,19 @@
   "type": "module",
   "scripts": {
     "dev": "ts-node-dev --quiet --project tsconfig.node.json src/dev.ts",
-    "build": "svelte-kit build && tsc --project tsconfig.node.json",
+    "build": "vite build && tsc --project tsconfig.node.json",
     "prepack": "./build-tools/replace_package_type.sh module commonjs",
     "postpack": "./build-tools/replace_package_type.sh commonjs module",
     "postinstall": "node build-tools/post_install.cjs",
     "postinstall:dev": "ts-node-dev --quiet --project tsconfig.node.json src/cli/post-install.ts",
     "postinstall:prod": "node dist/cli/post-install.js",
     "install-and-build": "npm install && npm run build",
-    "package": "svelte-kit package",
-    "preview": "svelte-kit preview",
     "check": "svelte-check --tsconfig ./tsconfig.json",
     "check:watch": "svelte-check --tsconfig ./tsconfig.json --watch",
     "lint": "eslint --ignore-path .gitignore .",
     "format": "prettier --ignore-path .gitignore --write --plugin-search-dir=. .",
     "duckdb": "node ./server/server.duckdb.js",
-    "dev:ui": "svelte-kit dev",
+    "dev:ui": "vite dev",
     "dev:backend": "ts-node-dev --project tsconfig.node.json -- src/server/server.ts",
     "test": "npm run test:backend && npm run test:ui",
     "test:backend": "jest --testPathIgnorePatterns test/ui",
@@ -37,9 +35,8 @@
     "@playwright/test": "^1.20.1",
     "@reduxjs/toolkit": "^1.8.2",
     "@rollup/plugin-typescript": "^8.3.1",
-    "@sveltejs/adapter-auto": "next",
-    "@sveltejs/adapter-static": "^1.0.0-next.34",
-    "@sveltejs/kit": "^1.0.0-next.355",
+    "@sveltejs/adapter-static": "^1.0.0-next.41",
+    "@sveltejs/kit": "^1.0.0-next.463",
     "@sveltestack/svelte-query": "^1.6.0",
     "@testing-library/jest-dom": "^5.16.3",
     "@testing-library/svelte": "^3.1.0",
