+++ src/routes/+layout.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { browser } from "$app/env";
+  import { browser } from "$app/environment";
   import { createStore } from "$lib/application-state-stores/application-store";
   import {
     createDerivedModelStore,
